// packages/core/src/types/view-computed.d.ts
export interface ComputedBranchCollectionPath<A extends AnyAux = AnyAux> {
    pathId: string;
    pathIndex: number;
    pathName?: string;
    pathTitle?: string | null;
    description?: scalar.MarkdownOrString;
    tags?: aux.Tags<A>;
    /**
     * Ordered list of edge ids emitted for this path.
     */
    edgeIds: readonly scalar.EdgeId[];
    isDefaultPath?: boolean;
}
export interface ComputedBranchCollection<A extends AnyAux = AnyAux> {
    branchId: string;
    astPath: string;
    kind: 'parallel' | 'alternate';
    label?: string;
    defaultPathId?: string;
    paths: readonly ComputedBranchCollectionPath<A>[];
}